Move indexers to independent module
